Function for sorting articles (määräykset) by huomioitava koodi created.
Lists of section 5 are sorted by huomioitava koodi. This applies to page 1 and two of KJ wizard.
